# Use Eigen::logistic instead of manually computing values.