**Description**: 
Use Eigen::logistic instead of ComputeLogistic for LinearClassifier

**Motivation and Context**
Close performance gap with SKL due to ComputeLogistic being slow.

Original changes had Eigen::logistic.

Batch size | v1.1.2 | master + Eigen::logistic | Speed-up
-- | -- | -- | --
100 | 0.0142334 | 0.0102302 | 1.39
1000 | 0.0971161 | 0.0385092 | 2.52
10000 | 0.884861 | 0.392591 | 2.25
100000 | 8.89221 | 3.50334 | 2.54

Latest changes have MlasComputeLogistic

Batch size | v1.1.2 | master + MlasComputeLogistic | Speed-up
-- | -- | -- | --
100 | 0.0141228 | 0.00986273 | 1.43
1000 | 0.0925429 | 0.0345273 | 2.68
10000 | 0.867377 | 0.351056 | 2.47
100000 | 8.56031 | 3.1005 | 2.76