@skottmckay 2020-03-12
Tracy just made me aware of MlasComputeLogistic. Will test out that and use it instead if possible.