@skottmckay 2020-03-12
The Linux builds are failing due to some tiny diffs in the output from the logreg_iris.onnx test model. The diffs are about the same magnitude on Windows, so maybe python unittest.TestCase.assertEqual has a slightly different tolerance on Linux. The diffs seem acceptable ('abs(expected - actual)' below) so I'll look into whether there's some configuration setting that can be used. 

Linux | Windows
-- | --
5.96046E-08 | 5.96046E-08
2.6077E-08 | 2.6077E-08
3.1665E-08 | 3.1665E-08
0 | 0
2.97587E-08 | 2.97587E-08
2.95695E-08 | 2.95695E-08
0 | 0
5.95903E-08 | 5.95903E-08
2.91038E-11 | 2.91038E-11